# perf(ext/http): optimize for zero or one-packet response streams