--- a/cli/tests/unit/serve_test.ts
+++ b/cli/tests/unit/serve_test.ts
@@ -538,11 +538,9 @@ function createStreamTest(count: number, delay: number, action: string) {
       if (action == "Throw") {
         controller.error(new Error("Expected error!"));
       } else {
-        console.log("close");
         controller.close();
       }
     } else {
-      console.log(i);
       controller.enqueue(`a${i}`);
 
       if (delay == 0) {
@@ -580,15 +578,22 @@ function createStreamTest(count: number, delay: number, action: string) {
 
     await listeningPromise;
     const resp = await fetch("http://127.0.0.1:4501/");
+    const text = await resp.text();
 
     ac.abort();
     await server;
     let expected = "";
-    for (let i = 0; i < count; i++) {
-      expected += `${i}`;
+    if (action == "Throw" && count < 2 && delay < 1000) {
+      // NOTE: This is specific to the current implementation. In some cases where a stream errors, we
+      // don't send the first packet.
+      expected = "";
+    } else {
+      for (let i = 0; i < count; i++) {
+        expected += `a${i}`;
+      }
     }
 
-    assertEquals(await resp.text(), expected);
+    assertEquals(text, expected);
   });
 }