This should improve `deno_reactdom_ssr_flash.jsx`: 39k req/s -> 63k req/s


```
09:29 $ third_party/prebuilt/mac/wrk http://localhost:4500
Running 10s test @ http://localhost:4500
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   251.00us   75.18us   4.43ms   90.92%
    Req/Sec    19.81k   773.52    20.62k    97.52%
  398137 requests in 10.10s, 68.72MB read
Requests/sec:  39429.22
Transfer/sec:      6.81MB

09:52 $ third_party/prebuilt/mac/wrk http://localhost:4500
Running 10s test @ http://localhost:4500
  2 threads and 10 connections
  Thread Stats   Avg      Stdev     Max   +/- Stdev
    Latency   155.35us   56.63us   3.84ms   91.88%
    Req/Sec    31.71k     1.48k   33.45k    87.62%
  637327 requests in 10.10s, 98.46MB read
Requests/sec:  63103.75
Transfer/sec:      9.75MB
```